+++ src/components/aside-news/index.module.scss
@@ -15,24 +15,41 @@
     line-height: 1;
     font-weight: 400;
     margin-bottom: 17px;
+    @media (max-width: 768px) {
+      font-size: 35px;
+      margin-bottom: 25px;
+    }
   }
 
   li {
     line-height: 1.1;
-
+    display: flex;
+    flex-direction: column;
+    gap: 5px;
     &:not(:last-child) {
       margin-bottom: 22px;
+      @media (max-width: 768px) {
+        margin-bottom: 31px;
+      }
     }
 
     span {
       font-size: 14px;
       color: #9D9D9D;
       display: block;
+      @media (max-width: 768px) {
+        font-size: 15px;
+        color: #ACACAC;
+      }
     }
 
     a {
       font-size: 14px;
       color: #015DB9;
+      @media (max-width: 768px) {
+        font-size: 18px;
+        color: #DE0008;
+      }
     }
   }
