+++ src/pages/news-page/layout/news-details/index.module.scss
@@ -1,79 +1,100 @@
-.newsItemPage {
-  display: flex;
-  gap: 20px;
-  justify-content: space-between;
-  @media (max-width: 768px) {
-    flex-direction: column;
-  }
-
-  .newsItemPageContent {
-    flex-grow: 1;
-    display: flex;
-    flex-direction: column;
-    @media (max-width: 768px) {
-      margin-bottom: 0;
-    }
-
+.newsContainer {
+  .newsListPage {
     h2 {
-      line-height: 1.1;
-      margin-bottom: 35px;
+      margin-bottom: 25px;
     }
-  }
-}
+    .newsItemPage {
+      display: flex;
+      gap: 20px;
+      justify-content: space-between;
+      @media (max-width: 768px) {
+        flex-direction: column;
+      }
 
+      .newsItemPageContent {
+        flex-grow: 1;
+        display: flex;
+        flex-direction: column;
+        @media (max-width: 768px) {
+          margin-bottom: 0;
+        }
 
-.newsItemDate {
-  color: #DE0008;
-  display: block;
-  margin-bottom: 30px;
-}
-
-.newsItemMainImg {
-  width: 100%;
-  height: 492px;
-  margin-bottom: 30px;
-  @media (max-width: 768px) {
-    height: 200px;
-  }
-
-  img {
-    width: 100%;
-    height: 100%;
-  }
-}
-
-.newsDescs {
-  p {
-    font-size: 16px;
-    &:not(:last-child) {
-      margin-bottom: 15px;
+        h2 {
+          line-height: 1.1;
+          margin-bottom: 35px;
+        }
+      }
     }
-  }
-}
 
-.newsGallery {
-  margin-top: 40px;
-  margin-bottom: 40px;
-  gap: 15px;
-  grid-template-columns: repeat(auto-fit, minmax(208px, 208px));
-  li {
-    width: 208px;
-    max-width: 208px;
-    & > div {
-      border: none;
-      height: 156px;
+
+    .newsItemDate {
+      color: #DE0008;
+      display: block;
+      margin-bottom: 30px;
+      font-size: 18px;
     }
-  }
-}
 
-.allNewsBlock {
-  margin-top: auto;
-  padding-top: 30px;
-  border-top: 7px solid #F2F5F8;
+    .newsItemMainImg {
+      width: 100%;
+      height: 492px;
+      margin-bottom: 30px;
+      @media (max-width: 768px) {
+        height: 173px;
+      }
 
-  a {
-    font-size: 16px;
-    color: #DE0008;
-    font-weight: 400;
+      img {
+        width: 100%;
+        height: 100%;
+      }
+    }
+
+    .newsDescs {
+      p {
+        font-size: 18px;
+        &:not(:last-child) {
+          margin-bottom: 15px;
+        }
+      }
+    }
+
+    .newsGallery {
+      margin-top: 40px;
+      margin-bottom: 40px;
+      gap: 15px;
+      grid-template-columns: repeat(auto-fit, minmax(208px, 208px));
+      li {
+        width: 208px;
+        max-width: 208px;
+        & > div {
+          border: none;
+          height: 156px;
+        }
+      }
+    }
+
+    .gallerySlider {
+      margin: 30px 0 50px 0;
+      .galleryBtns {
+        @media (max-width: 768px) {
+          top: 50%;
+        }
+      }
+    }
+
+    .allNewsBlock {
+      margin-top: auto;
+      padding-top: 30px;
+      border-top: 7px solid #F2F5F8;
+
+      a {
+        font-size: 18px;
+        color: #DE0008;
+        font-weight: 400;
+      }
+    }
+    @media (max-width: 768px) {
+      min-height: auto;
+      margin: 0 0 45px 0;
+    }
   }
 }
