--- src/pages/news-page/layout/news/components/filtered-news-list/index.module.scss
@@ -1,13 +1,0 @@
-.newsList {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(270px, 270px));
-  gap: 23px;
-}
-
-.newsAbsence {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  padding-top: 150px;
-  font-size: 25px;
-}
