+++ src/pages/news-page/news-layout.tsx
@@ -5,16 +5,18 @@ import { Outlet } from 'react-router-dom'
 
 export const NewsLayout: FC = () => {
 	return (
-		<Container>
-			<BreadCrumbs
-				crumbsLinksMap={[
-					{
-						title: 'Все новости',
-						link: 'news',
-					},
-				]}
-			/>
+		<>
+			<Container>
+				<BreadCrumbs
+					crumbsLinksMap={[
+						{
+							title: 'Все новости',
+							link: 'news',
+						},
+					]}
+				/>
+			</Container>
 			<Outlet />
-		</Container>
+		</>
 	)
 }
